working out client logger issues

+++ tinker_access_client/tinker_access_client/ServiceInstaller.py
@@ -33,16 +33,17 @@ class ServiceInstaller(object):
 
     # TODO: this can be removed after the upgrade
     def __remove_legacy_client(self):
-        self.__execute_commands([
-            'service tinkerclient stop\n',
-            'update-rc.d -f tinkerclient remove\n',
-            'rm -rf /etc/init.d/tinkerclient\n',
-            'rm -rf /opt/tinkeraccess/client.py\n',
-            'rm -rf /opt/tinkeraccess/client.pyc\n',
-            'rm -rf /opt/tinkeraccess/lcdModule.py\n',
-            'rm -rf /opt/tinkeraccess/lcdModule.pyc\n',
-            'rm -rf /opt/tinkeraccess/client.cfg\n'
-        ])
+        if os.path.exists('/etc/init.d/tinkerclient'):
+            self.__execute_commands([
+                'service tinkerclient stop\n',
+                'update-rc.d -f tinkerclient remove\n',
+                'rm -rf /etc/init.d/tinkerclient\n',
+                'rm -rf /opt/tinkeraccess/client.py\n',
+                'rm -rf /opt/tinkeraccess/client.pyc\n',
+                'rm -rf /opt/tinkeraccess/lcdModule.py\n',
+                'rm -rf /opt/tinkeraccess/lcdModule.pyc\n',
+                'rm -rf /opt/tinkeraccess/client.cfg\n'
+            ])
 
     def __create_service(self):
         self.__ensure_execute_permission(self.__service_script)
